requirements.txt
openai>=1.0.0
httpx>=0.27
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
pydantic>=2.0
chainlit>=2.0
